+++ app/src/main/java/me/stageguard/aruku/database/message/MessageRecordEntity.kt
@@ -31,7 +31,7 @@ data class MessageRecordEntity(
     @Embedded val contact: ContactId,
     @ColumnInfo(name = "sender") val sender: Long,
     @ColumnInfo(name = "sender_name") val senderName: String,
-    @PrimaryKey @ColumnInfo(name = "message_id") val messageId: Int,
+    @PrimaryKey @ColumnInfo(name = "message_id") val messageId: Long,
     @ColumnInfo(name = "time") val time: Long,
     @ColumnInfo(name = "message") val message: List<MessageElement>,
 ) : Parcelable
