+++ app/src/main/java/me/stageguard/aruku/ui/page/chat/ChatViewModel.kt
@@ -14,8 +14,8 @@ import kotlinx.coroutines.flow.*
 import me.stageguard.aruku.cache.AudioCache
 import me.stageguard.aruku.domain.MainRepository
 import me.stageguard.aruku.domain.data.message.*
-import me.stageguard.aruku.service.parcel.ContactType
 import me.stageguard.aruku.service.bridge.AudioStatusListener
+import me.stageguard.aruku.service.parcel.ContactType
 import me.stageguard.aruku.ui.UiState
 import me.stageguard.aruku.ui.page.ChatPageNav
 import me.stageguard.aruku.util.toFormattedTime
@@ -147,7 +147,7 @@ sealed interface ChatElement {
         val senderName: String,
         val senderAvatarUrl: String?,
         val time: String,
-        val messageId: Int,
+        val messageId: Long,
         val visibleMessages: List<VisibleChatMessage>
     ) : ChatElement {
         override val uniqueKey = messageId
